refactor: 重构bit操作方法参数为uint8_t

+++ src/bits.c
@@ -4,11 +4,11 @@ bool bit_get(uint8_t value, uint8_t index) {
     return BIT_GET(value, index);
 }
 
-void bit_set(uint8_t* value, int index, bool state) {
+void bit_set(uint8_t* value, uint8_t index, bool state) {
     BIT_SET(value, index, state);
 }
 
-uint8_t bit_set_new(uint8_t value, int index, bool state) {
+uint8_t bit_set_new(uint8_t value, uint8_t index, bool state) {
     if (state) {
         return value | (1UL << index);
     } else {
